mount: change check from stat to `mountpoint`
Again, as in the previous commit, we may have a situation where the fs
wasn't unmounted cleanly. If we just stat, the directory may still exist
and we won't mount it, and things will be wrong.

Instead, let's parse mountinfo, and check explicitly to see if all of these
are a mountpoint, and mount them if they're not.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>