add a way to mount OCI stuff without DB interaction
We're interested in using the mount code without having to interact with
the DB at all. This commit does a few things:

* refactor Config to be an interface
* makes the original Config a DBBasedConfig
* adds a new OCI type config that doesn't interact with the DB

Eventually we should add a cli interface that mounts stuff directly from
OCI, but we leave that for future work.

Note that we still need a "metadata" spot to put the mounts and writable
areas, etc. Worth noting though that this can just be scratch space on a
tpmfs or whatever, unless you want to persist the writable stuff.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>