deps: bump stacker, umoci, and mtree versions
We can get rid of the mtree replace because there is a new mtree release
with the features that we need

We need to bump umoci, because it moved.

We need to bump stacker, because we use its OCI API wrappers, and their
types have changed to the new path.

Signed-off-by: Tycho Andersen <tycho@tycho.ws>